{"types":"C","runs":23}

+++ Problem22.java
@@ -4,7 +4,7 @@ import acm.program.GraphicsProgram;
 public class Problem22 extends GraphicsProgram {
 	private static final int NUM_CELLS = 8;
 	public void run() {
-		double squareLength = Math.min(getWidth() / NUM_CELLS, getHeight() / NUM_CELLS);
+		double squareLength = Math.min(getWidth() / NUM_CELLS, (double)getHeight() / NUM_CELLS);
 		double x = 0;
 		double y = 0;
 		for(int j = 0; j < NUM_CELLS; j++) {
